refactor: move prayResponseDto response package
