+++ src/main/java/com/uspray/uspray/DTO/sharedpray/response/SharedPrayListResponseDto.java
@@ -1,7 +1,6 @@
 package com.uspray.uspray.DTO.sharedpray.response;
 
-import com.uspray.uspray.DTO.pray.request.PrayResponseDto;
-import com.uspray.uspray.domain.SharedPray;
+import com.uspray.uspray.DTO.pray.response.PrayResponseDto;
 import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
